Title: 사칙연산을 이행하는 함수, Time: 117ms, Memory: 27MB, Status: Passed - CodeTree

+++ 240802/사칙연산을 이행하는 함수/functions-that-carry-out-four-factor-operations.py
@@ -8,6 +8,6 @@ elif arr[1] == '-':
 elif arr[1] == '*':
     result = x * y
 else:
-    result = x / y
+    result = int(x / y)
 
 print(f'{x} {arr[1]} {y} = {result}')
